+++ src/kernel/SparseGaussianKernel.cpp
@@ -7,7 +7,7 @@
 
 #include <assert.h>
 
-CSparseGaussianKernel::CSparseGaussianKernel(long size, double w)
+CSparseGaussianKernel::CSparseGaussianKernel(LONG size, double w)
   : CSparseRealKernel(size),width(w)
 {
 }
@@ -37,9 +37,9 @@ bool CSparseGaussianKernel::save_init(FILE* dest)
 	return false;
 }
   
-REAL CSparseGaussianKernel::compute(long idx_a, long idx_b)
+REAL CSparseGaussianKernel::compute(INT idx_a, INT idx_b)
 {
-  long alen, blen;
+  INT alen, blen;
   bool afree, bfree;
 
   TSparseEntry<REAL>* avec=((CSparseRealFeatures*) lhs)->get_sparse_feature_vector(idx_a, alen, afree);
@@ -47,7 +47,7 @@ REAL CSparseGaussianKernel::compute(long idx_a, long idx_b)
   
   REAL result=0;
 
-  long i;
+  INT i;
   for (i=0; i<alen; i++)
 	  result+= avec[i].entry * avec[i].entry;
 
@@ -57,10 +57,10 @@ REAL CSparseGaussianKernel::compute(long idx_a, long idx_b)
 
   if (alen<=blen)
   {
-	  long j=0;
-	  for (long i=0; i<alen; i++)
+	  INT j=0;
+	  for (INT i=0; i<alen; i++)
 	  {
-		  int a_feat_idx=avec[i].feat_index;
+		  INT a_feat_idx=avec[i].feat_index;
 
 		  while ( (j<blen) && (bvec[j].feat_index < a_feat_idx) )
 			  j++;
@@ -74,10 +74,10 @@ REAL CSparseGaussianKernel::compute(long idx_a, long idx_b)
   }
   else
   {
-	  long j=0;
-	  for (long i=0; i<blen; i++)
+	  INT j=0;
+	  for (INT i=0; i<blen; i++)
 	  {
-		  int b_feat_idx=bvec[i].feat_index;
+		  INT b_feat_idx=bvec[i].feat_index;
 
 		  while ( (j<alen) && (avec[j].feat_index < b_feat_idx) )
 			  j++;
