+++ src/preproc/NormOne.cpp
@@ -43,9 +43,9 @@ bool CNormOne::save(FILE* f)
 /// return pointer to feature_matrix, i.e. f->get_feature_matrix();
 REAL* CNormOne::apply_to_feature_matrix(CFeatures* f)
 {
-	long i,j;
-	long num_vec;
-	long num_feat;
+	INT i,j;
+	INT num_vec;
+	INT num_feat;
 	REAL* matrix=((CRealFeatures*) f)->get_feature_matrix(num_feat, num_vec);
 
 	for (i=0; i<num_vec; i++)
@@ -71,12 +71,12 @@ REAL* CNormOne::apply_to_feature_matrix(CFeatures* f)
 
 /// apply preproc on single feature vector
 /// result in feature matrix
-REAL* CNormOne::apply_to_feature_vector(REAL* f, int& len)
+REAL* CNormOne::apply_to_feature_vector(REAL* f, INT& len)
 {
 	REAL* vec=new REAL[len];
 	REAL sqnorm=0;
 	REAL norm=0;
-	long i=0;
+	INT i=0;
 
 	for (i=0; i<len; i++)
 		sqnorm+=f[i]*f[i];
